+++ src/lib/pages/account-details/components/tables/AdminContractsTable.tsx
@@ -2,16 +2,18 @@ import { Box } from "@chakra-ui/react";
 import type { ChangeEvent } from "react";
 
 import { ErrorFetching } from "../ErrorFetching";
-import { MobileTitle } from "../mobile/MobileTitle";
 import { useInternalNavigate, useMobile } from "lib/app-provider";
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
 import { EmptyState } from "lib/components/state";
-import { TableTitle, ViewMore } from "lib/components/table";
+import {
+  ContractsTable,
+  MobileTitle,
+  TableTitle,
+  ViewMore,
+} from "lib/components/table";
 import { useAccountAdminContracts } from "lib/pages/account-details/data";
 import type { ContractAddr, HumanAddr, Option } from "lib/types";
-
-import { ContractsBody } from "./ContractsBody";
 
 interface AdminContractsTableProps {
   walletAddress: HumanAddr;
@@ -21,31 +23,6 @@ interface AdminContractsTableProps {
   onViewMore?: () => void;
 }
 
-const AdminTitle = ({
-  onViewMore,
-  totalData,
-}: {
-  onViewMore: AdminContractsTableProps["onViewMore"];
-  totalData: Option<number>;
-}) => {
-  const isMobile = useMobile();
-  if (isMobile && onViewMore)
-    return (
-      <MobileTitle
-        title="Contract Admins"
-        count={totalData ?? 0}
-        onViewMore={onViewMore}
-      />
-    );
-  return (
-    <TableTitle
-      title="Contract Admins"
-      count={totalData ?? 0}
-      helperText="This account is the admin for following contracts"
-      mb={2}
-    />
-  );
-};
 export const AdminContractsTable = ({
   walletAddress,
   scrollComponentId,
@@ -53,6 +30,7 @@ export const AdminContractsTable = ({
   refetchCount,
   onViewMore,
 }: AdminContractsTableProps) => {
+  const isMobile = useMobile();
   const navigate = useInternalNavigate();
   const onRowSelect = (contract: ContractAddr) =>
     navigate({
@@ -92,35 +70,42 @@ export const AdminContractsTable = ({
     setPageSize(size);
     setCurrentPage(1);
   };
-  const isMobile = useMobile();
-  let isMobileDetail = null;
-  if (isMobile && onViewMore) {
-    isMobileDetail = false;
-  } else {
-    isMobileDetail = true;
-  }
+
+  const isMobileOverview = isMobile && !!onViewMore;
   return (
     <Box mt={{ base: 4, md: 8 }}>
-      <AdminTitle totalData={totalData} onViewMore={onViewMore} />
-      {isMobileDetail && (
-        <ContractsBody
-          contracts={contracts}
-          isLoading={isLoading}
-          emptyState={
-            <EmptyState
-              message={
-                !contracts ? (
-                  <ErrorFetching />
-                ) : (
-                  "This account does not have any admin access for any contracts."
-                )
-              }
-              withBorder
-            />
-          }
-          onRowSelect={onRowSelect}
+      {isMobileOverview ? (
+        <MobileTitle
+          title="Contract Admins"
+          count={totalData ?? 0}
           onViewMore={onViewMore}
         />
+      ) : (
+        <>
+          <TableTitle
+            title="Contract Admins"
+            count={totalData ?? 0}
+            helperText="This account is the admin for following contracts"
+            mb={2}
+          />
+          <ContractsTable
+            contracts={contracts}
+            isLoading={isLoading}
+            emptyState={
+              <EmptyState
+                message={
+                  !contracts ? (
+                    <ErrorFetching />
+                  ) : (
+                    "This account does not have any admin access for any contracts."
+                  )
+                }
+                withBorder
+              />
+            }
+            onRowSelect={onRowSelect}
+          />
+        </>
       )}
       {!!totalData &&
         (onViewMore