+++ src/lib/pages/module-details/components/ModuleStruct.tsx
@@ -4,7 +4,7 @@ import { useMemo, useState } from "react";
 import { CustomIcon } from "lib/components/icon";
 import InputWithIcon from "lib/components/InputWithIcon";
 import { StructCard } from "lib/components/module/StructCard";
-import type { IndexedModule } from "lib/services/moduleService";
+import type { IndexedModule } from "lib/services/move/moduleService";
 
 interface ModuleStructProps {
   structs: IndexedModule["parsedAbi"]["structs"];
@@ -15,7 +15,9 @@ export const ModuleStruct = ({ structs }: ModuleStructProps) => {
 
   const filteredStructs = useMemo(() => {
     if (!keyword) return structs;
-    return structs.filter((struct) => struct.name.includes(keyword));
+    return structs.filter((struct) =>
+      struct.name.toLowerCase().includes(keyword.toLowerCase())
+    );
   }, [keyword, structs]);
 
   const updateExpandedIndexes = (indexes: number[]) =>